Refactor: se reestructura el proyecto entero
