--- src/main/java/com/gupoti/mexico/portalarrendamiento/Rest/C_ConfigurationRest.java
@@ -1,7 +1,0 @@
-package com.gupoti.mexico.portalarrendamiento.Rest;
-
-public class C_ConfigurationRest {
-
-
-
-}
