--- src/main/java/com/gupoti/mexico/portalarrendamiento/Service/C_ConfigurationServ.java
@@ -1,7 +1,0 @@
-package com.gupoti.mexico.portalarrendamiento.Service;
-
-
-public class C_ConfigurationServ {
-	
-
-}
