+++ src/main/java/com/gupoti/mexico/portalarrendamiento/Service/Catalogos/ConfigurationServ.java
@@ -1,0 +1,6 @@
+package com.gupoti.mexico.portalarrendamiento.Service.Catalogos;
+
+
+public interface ConfigurationServ {
+
+}
