+++ src/main/resources/application.properties
@@ -1,5 +1,4 @@
 spring.application.name=PortArrendamientoBack
-
 spring.datasource.url=jdbc:postgresql:${URL}
 spring.datasource.username=${USERNAME}
 spring.datasource.password=${PASSWORD}
